+++ restaurant/src/front_of_house/mod.rs
@@ -1,0 +1,3 @@
+pub mod hosting;
+
+mod serving;
