+++ restaurant/src/front_of_house/serving.rs
@@ -1,0 +1,5 @@
+fn take_order() {}
+
+fn serve_order() {}
+
+fn take_payment() {}
