+++ restaurant/src/lib.rs
@@ -1,21 +1,4 @@
-mod front_of_house {
-
-    pub mod hosting {
-
-        pub fn add_to_waitlist() {}
-
-        fn seat_at_table() {}
-    }
-
-    mod serving {
-
-        fn take_order() {}
-
-        fn serve_order() {}
-
-        fn take_payment() {}
-    }
-}
+mod front_of_house;
 
 // the use of `pub` keyword here denotes it is being re-exported.
 // meaning the caller module can reference simply `restaurant::hosting::add_to_waitlist()`
@@ -46,14 +29,13 @@ pub fn eat_at_restaurant() {
     let order2 = back_of_house::Appetizer::Salad;
 }
 
-
 fn serve_order() {}
 
 mod back_of_house {
-    
+
     fn fix_incorrect_order() {
         cook_order();
-        super::serve_order();        
+        super::serve_order();
     }
 
     fn cook_order() {}
